+++ docs/base/release_notes.rst
@@ -17,6 +17,8 @@ devices have been selected.
 - Make Netmiko prompt command service a substitution string in the UI
 - Fix wrong jump password when using a Vault
 - Fix workflow results recursive display no path in results bug
+- Improve "Get Result" REST endpoint: returns 404 error if no run found, run status if a run is found but there are
+no results (e.g job still running), and the results if the job is done.
 
 Version 3.21.2
 --------------